feat : separate feature to component #39
