+++ src/components/Home/Main/style.ts
@@ -35,70 +35,3 @@ export const Wrap = styled.div`
   gap: 16px;
   padding: 24px;
 `;
-
-export const InfoContainer = styled.div`
-  width: 51%;
-  & > :nth-child(2) {
-    display: none;
-  }
-  @media (max-width: 1068px) {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    & > :nth-child(2) {
-    display: block;
-    }
-  }
-`
-
-export const ApplyInfoContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  @media (max-width: 1068px) {
-    & > :nth-child(2) {
-    display: none;
-    }
-  }
-`
-
-export const ApplyInfo = styled.div`
-  width: 100%;
-
-  display: flex;
-  gap: 12px;
-  & > :nth-child(2) {
-    display: none;
-  }
-  @media (max-width: 1068px) {
-    display: flex;
-    flex-direction: column;
-    & > :nth-child(2) {
-    display: block;
-    }
-  }
-`;
-
-export const Indicator = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  gap: 12px;
-`
-
-export const IndicatorButton = styled.div<{ $focus : "true" | "false" }>`
-  background-color: ${({ $focus, theme }) => $focus === "true" ? theme.primaryNormal : theme.labelDisabled };
-  border-radius: 99px;
-  width: 10px;
-  height: 10px;
-  cursor: pointer;
-`
-
-export const ButtonWrap = styled.div`
-  width: 100%;
-
-  display: flex;
-  justify-content: flex-end;
-
-  padding-right: 10px;
-`;
